Perf fix

+++ src/core/alex_base.h
@@ -256,7 +256,7 @@ class ExpectedShiftsAccumulator : public StatAccumulator {
   // Therefore, we track n^2/4.
   void accumulate(int actual_position, int) override {
     if (actual_position > last_position_ + 1) {
-      int dense_region_length = last_position_ - dense_region_start_idx_ + 1;
+      long long dense_region_length = last_position_ - dense_region_start_idx_ + 1;
       num_expected_shifts_ += (dense_region_length * dense_region_length) / 4;
       dense_region_start_idx_ = actual_position;
     }
@@ -267,7 +267,7 @@ class ExpectedShiftsAccumulator : public StatAccumulator {
   double get_stat() override {
     if (count_ == 0) return 0;
     // first need to accumulate statistics for current packed region
-    int dense_region_length = last_position_ - dense_region_start_idx_ + 1;
+    long long dense_region_length = last_position_ - dense_region_start_idx_ + 1;
     long long cur_num_expected_shifts =
         num_expected_shifts_ + (dense_region_length * dense_region_length) / 4;
     return cur_num_expected_shifts / static_cast<double>(count_);
@@ -300,7 +300,7 @@ class ExpectedIterationsAndShiftsAccumulator : public StatAccumulator {
         std::log2(std::abs(predicted_position - actual_position) + 1);
 
     if (actual_position > last_position_ + 1) {
-      int dense_region_length = last_position_ - dense_region_start_idx_ + 1;
+      long long dense_region_length = last_position_ - dense_region_start_idx_ + 1;
       num_expected_shifts_ += (dense_region_length * dense_region_length) / 4;
       dense_region_start_idx_ = actual_position;
     }
@@ -321,7 +321,7 @@ class ExpectedIterationsAndShiftsAccumulator : public StatAccumulator {
 
   double get_expected_num_shifts() {
     if (count_ == 0) return 0;
-    int dense_region_length = last_position_ - dense_region_start_idx_ + 1;
+    long long dense_region_length = last_position_ - dense_region_start_idx_ + 1;
     long long cur_num_expected_shifts =
         num_expected_shifts_ + (dense_region_length * dense_region_length) / 4;
     return cur_num_expected_shifts / static_cast<double>(count_);
